refactor: Rename cluster to modal
